Ensure the "confidential" property is set
If not confidential, then it is possible that the "confidential" property is not set and the default of false is used. This fails a filter test for (confidential=*) since it is then not set. This makes it very hard to find a filter that matches both a confidential env. and a non-confidential env, e.g.: (|(confidential=true)(port=*)). This change just sets confidential to false in that case so (confidential=*) works.